rename files and remove debug log
