+++ src/js/seqMsgParser.mjs
